Split OpenAI payload adapters
Move chat completion, Responses, and shared validation logic into
focused modules under `episodic.llm` while keeping `openai_client` as
the compatibility import surface.
